sentiment_classification/requirements.txt
pytorch
deepspeed
gensim
scikit-learn
numpy